Add database hook for contact saves
Database helper is now hooked up to the contact selection functionality. The contacts are freely changeable on the course of the activity and will only be queried to the db when the save button is hit.

TODO: Implement saved states where app will repopulated app data if data was filled and saved before