Add app restore if contacts previously saved
The app will now restore data that has been saved to the database. As long as the user inputs contacts and presses “save”, the next time the app opens (in any state), the data will be pre-populated. That is, I did not use saved instance state in order to save and restore our data. It is done using a query to the database and simply going through the data and filling it in since it’s a small dataset.